requirements.txt
opencv-python==4.10.0.84
mediapipe==0.10.14
pyserial==3.5
